backend/tests/requirements.txt
psycopg2-binary~=2.9.10
pytest~=8.3.5
requests~=2.32.3